Move Tilesets out of ui.py
