+++ cddagl/ui/views/tilesets.py
@@ -1,0 +1,22 @@
+import logging
+
+from PyQt5.QtWidgets import QTabWidget
+
+import cddagl
+
+version = cddagl.__version__
+logger = logging.getLogger('cddagl')
+
+
+class TilesetsTab(QTabWidget):
+    def __init__(self):
+        super(TilesetsTab, self).__init__()
+
+    def set_text(self):
+        pass
+
+    def get_main_window(self):
+        return self.parentWidget().parentWidget().parentWidget()
+
+    def get_main_tab(self):
+        return self.parentWidget().parentWidget().main_tab
